update abm section on integration page
