+++ frontend/pages/admin/IntegrationsPage/cards/MdmSettings/components/AppleBusinessManagerSection/_styles.scss
@@ -1,4 +1,4 @@
-.automatic-enrollment-section {
+.apple-business-manager-section {
   &__content {
     display: flex;
     flex-direction: column;
