+++ frontend/pages/admin/IntegrationsPage/cards/MdmSettings/components/AppleBusinessManagerSection/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./AppleBusinessManagerSection";
