--- frontend/pages/admin/IntegrationsPage/cards/MdmSettings/components/AutomaticEnrollmentSection/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./AutomaticEnrollmentSection";
