--- frontend/pages/admin/IntegrationsPage/cards/MdmSettings/components/VppSection/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./VppSection";
