+++ frontend/pages/admin/IntegrationsPage/cards/MdmSettings/components/WindowsAutopilotSection/WindowsAutomaticEnrollmentCard/WindowsAutomaticEnrollmentCard.tsx
