+++ frontend/pages/admin/IntegrationsPage/cards/MdmSettings/components/WindowsAutopilotSection/WindowsAutomaticEnrollmentCard/index.ts
