+++ frontend/pages/admin/IntegrationsPage/cards/MdmSettings/components/WindowsAutopilotSection/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./WindowsAutopilotSection";
